+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 Change Log
 ===
 
+v2.0.1
+---
+* [enhancement] Added preliminary checks for `jasmine-no-lambda-expression-callbacks` rule to skip walking file if there are no jasmine
+top-level statements.
+
 v2.0.0
 ---
-* [enhancement] Upgrades to tslint@4
+* **BREAKING CHANGES**
+  * [enhancement] Upgraded to tslint@4
+
+v1.0.0
+---
+* [new-rule] `import-barrels` rule added
+* [new-rule] `jasmine-no-lambda-expression-callbacks` rule added